python/requirements.txt
beautifulsoup4
requests
pandas
lxml
pymongo
python-dotenv